ai/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain_huggingface
faiss-cpu
langchain-community
diskcache
jinja2
numpy
typing-extensions
llama-cpp-python
